History logic

+++ routes/index.py
@@ -61,7 +61,7 @@ def indexFunction():
             # Sum up the quantities
             newQuantity = existingQuantity + int(quantity)
             user_id = session["user_id"]
-            subtotal = int(newQuantity) * int(price)
+            subtotal = float(newQuantity) * float(price)
             
             # Update database with quantity
             query = Baskets.query.filter_by(plant_id=plant_id).filter_by(user_id=user_id).first()
@@ -74,7 +74,7 @@ def indexFunction():
         else:
 
             user_id = session["user_id"]
-            subtotal = int(quantity) * int(price)
+            subtotal = float(quantity) * float(price)
             
             # Add quantity to database
             db.session.add(Baskets(plant_id=plant_id, user_id=user_id, quantity=quantity, name=name, picture=picture, price=price, subtotal=subtotal))
